Enterprise-Knowledge-Assistant/services/fastapi-ingest/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pinecone-client==3.2.2
openai==1.58.0
httpx==0.27.0
pypdf==4.2.0
python-docx==1.1.0
Pillow==10.4.0
pytesseract==0.3.10
minio==7.2.7
prometheus-client==0.20.0